Database change, known issues update, Ubuntu notes change, minor spelling fixes
Adding note about changed database which requires to delete 'kisslib.db' once and recreate the launcher and reimport the ebooks. But the database might undergo still changes in the future to support other features.
Updated known issue, in case files cannot be imported after update, that the database file has to be recreated, this is a one time step.
Added note that Ubuntu 16.10 doesnt show the default GTK icons.
Minor spelling fixes.